+++ src/App.js
@@ -1,4 +1,4 @@
-import { Route, Routes } from "react-router-dom";
+import { Route, Routes, useNavigate } from "react-router-dom";
 import { useContext,useEffect } from "react";
 import "./App.css";
 import Header from "./components/Header/Header";
@@ -9,12 +9,58 @@ import HomePage from "./pages/HomePage/HomePage";
 import ProductListingPage from "./pages/ProductListingPage/ProductListingPage";
 import Signin from "./pages/Signin/Signin";
 import WishlistPage from "./pages/WishlistPage/WishlistPage";
+import AuthProvider, { AuthContext } from "./AuthProvider";
+import axios from "axios";
 
 function App() {
-  const encodedToken = localStorage.getItem('token')
+ 
+ 
   const {state,dispatch}=useContext(StateContext)
-  
+ const encodedToken = localStorage.getItem('token')
 
+  useEffect(() => {
+  
+    async function fetchData() {
+      try {
+        const res = await fetch('/api/user/cart', {
+          method: "GET", headers: {
+            "authorization": encodedToken,
+            "Content-type": "application/json; charset=UTF-8"
+          }
+        })
+        const data = await res.json()
+
+        dispatch({ type: 'SET_CART', payload: data.cart })
+
+      } catch (e) {
+        console.log(e)
+      }
+    };
+    fetchData();
+   
+  }, [state.cart])
+
+  useEffect(() => {
+     
+    async function fetchData() {
+      try {
+        const res = await fetch('/api/user/wishlist', {
+          method: "GET", headers: {
+            "authorization": encodedToken,
+            "Content-type": "application/json; charset=UTF-8"
+          }
+        })
+        const data = await res.json()
+
+        dispatch({ type: 'SET_WISHLIST', payload: data.wishlist })
+
+      } catch (e) {
+        console.log(e)
+      }
+    };
+    fetchData();
+  
+}, [state.wishlist])
   return (
     <>
       <TopBar />
