+++ src/AuthProvider.js
@@ -5,7 +5,7 @@ import { useState } from 'react';
 export const AuthContext = createContext();
 
 const AuthProvider = ({ children }) => {
-  const [auth, setAuth] = useState(localStorage.getItem("token"));
+  const [auth, setAuth] = useState((localStorage.getItem("token")));
   return (
     <AuthContext.Provider value={{ auth, setAuth }}>
       {children}
