+++ src/Context.js
@@ -7,10 +7,7 @@ import stateReducer from './Reducer';
 export const StateContext = createContext();
 
 const Context = ({ children }) => {
-
-  const encodedToken = localStorage.getItem('token')
  
-
   const [state, dispatch] = useReducer(stateReducer, {
     products: [],
     categories: [],
@@ -26,48 +23,11 @@ const Context = ({ children }) => {
     wishlist:[]
   })
 
-  useEffect(() => {
-    async function fetchData() {
-      try {
-        const res = await fetch('/api/user/cart', {
-          method: "GET", headers: {
-            "authorization": encodedToken,
-            "Content-type": "application/json; charset=UTF-8"
-          }
-        })
-        const data = await res.json()
-
-        dispatch({ type: 'SET_CART', payload: data.cart })
-
-      } catch (e) {
-        console.log(e)
-      }
-    };
-    fetchData();
-  }, [state.cart])
-
-  useEffect(() => {
-    async function fetchData() {
-      try {
-        const res = await fetch('/api/user/wishlist', {
-          method: "GET", headers: {
-            "authorization": encodedToken,
-            "Content-type": "application/json; charset=UTF-8"
-          }
-        })
-        const data = await res.json()
-
-        dispatch({ type: 'SET_WISHLIST', payload: data.wishlist })
-
-      } catch (e) {
-        console.log(e)
-      }
-    };
-    fetchData();
-  }, [state.wishlist])
+ 
 
 
-  return (
+
+   return (
     <StateContext.Provider value={{ state, dispatch }}>
       {children}
     </StateContext.Provider>
