+++ src/pages/WishlistPage/WishlistPage.js
@@ -1,11 +1,13 @@
 import React from 'react'
-import { useContext } from 'react';
+import { useContext,useEffect } from 'react';
 import ProductCard from '../../components/ProductCard/ProductCard';
 import { StateContext } from '../../Context';
 import './WishlistPage.css';
 
 const WishlistPage = () => {
     const {state,dispatch}=useContext(StateContext)
+    const encodedToken = localStorage.getItem('token')
+   
   return (
     <>
      <div class="product-display-container">
